refactor(projectDetails): break project details map component to another component
